requirements.txt
nicegui==3.9.0
fastapi==0.135.2
uvicorn[standard]>=0.27.0
pydantic>=2.0.0,<3.0
python-box>=7.0
requests>=2.31.0
wechat-work-webhook>=0.0.2
dingtalkchatbot>=1.5.7
schedule>=1.2.0
packaging>=23.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
argcomplete>=3.0.0
paramiko>=3.4.0
netaddr>=1.0.0
